refactor: move types into src directory
